Construct or fit an XGBoost Inducer.

Construct or fit an XGBoost Inducer.

InducerXGBoost(.data, ...)

Arguments

.data

Object of class Dataset. Optional.

...

Additional parameters specifying the hyperparameter configuration of the XGBoost model. If none are given, the default configuration of xgboost::xgboost is used for the model fitting process.

Value

A S3 object of either class Model (.data was given) or of class InducerXGBoost (.data was not given).

Details

If .data argument is passed, an XGBoost model will be trained on the given dataset using the passed hyperparameter configuration. Otherwise, an inducer of class InducerXGBoost is created with the given hyperparameter configuration.

Examples

InducerXGBoost()
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> NULL
#> 
#> $configuration
#> list()
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
InducerXGBoost(verbose = 0)
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "verbose"
#> 
#> $configuration
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
cars.data <- Dataset(data = cars, target = "dist")
InducerXGBoost(cars.data, verbose = 0, nrounds = 10)
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "verbose" "nrounds"
#> 
#> $configuration
#> $configuration$verbose
#> [1] 0
#> 
#> $configuration$nrounds
#> [1] 10
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
#> 
#> $data
#> $data
#>    dist speed
#> 1     2     4
#> 2    10     4
#> 3     4     7
#> 4    22     7
#> 5    16     8
#> 6    10     9
#> 7    18    10
#> 8    26    10
#> 9    34    10
#> 10   17    11
#> 11   28    11
#> 12   14    12
#> 13   20    12
#> 14   24    12
#> 15   28    12
#> 16   26    13
#> 17   34    13
#> 18   34    13
#> 19   46    13
#> 20   26    14
#> 21   36    14
#> 22   60    14
#> 23   80    14
#> 24   20    15
#> 25   26    15
#> 26   54    15
#> 27   32    16
#> 28   40    16
#> 29   32    17
#> 30   40    17
#> 31   50    17
#> 32   42    18
#> 33   56    18
#> 34   76    18
#> 35   84    18
#> 36   36    19
#> 37   46    19
#> 38   68    19
#> 39   32    20
#> 40   48    20
#> 41   52    20
#> 42   56    20
#> 43   64    20
#> 44   66    22
#> 45   54    23
#> 46   70    24
#> 47   92    24
#> 48   93    24
#> 49  120    24
#> 50   85    25
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 1     2     4
#> 2    10     4
#> 3     4     7
#> 4    22     7
#> 5    16     8
#> 6    10     9
#> 7    18    10
#> 8    26    10
#> 9    34    10
#> 10   17    11
#> 11   28    11
#> 12   14    12
#> 13   20    12
#> 14   24    12
#> 15   28    12
#> 16   26    13
#> 17   34    13
#> 18   34    13
#> 19   46    13
#> 20   26    14
#> 21   36    14
#> 22   60    14
#> 23   80    14
#> 24   20    15
#> 25   26    15
#> 26   54    15
#> 27   32    16
#> 28   40    16
#> 29   32    17
#> 30   40    17
#> 31   50    17
#> 32   42    18
#> 33   56    18
#> 34   76    18
#> 35   84    18
#> 36   36    19
#> 37   46    19
#> 38   68    19
#> 39   32    20
#> 40   48    20
#> 41   52    20
#> 42   56    20
#> 43   64    20
#> 44   66    22
#> 45   54    23
#> 46   70    24
#> 47   92    24
#> 48   93    24
#> 49  120    24
#> 50   85    25
#> 
#> $y
#>  [1]  4  4  7  7  8  9 10 10 10 11 11 12 12 12 12 13 13 13 13 14 14 14 14 15 15
#> [26] 15 16 16 17 17 17 18 18 18 18 19 19 19 20 20 20 20 20 22 23 24 24 24 24 25
#> 
#> $model
#> ##### xgb.Booster
#> raw: 16.1 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1   37.25719
#>         2   28.61702
#> ---                 
#>         9   12.29622
#>        10   12.03218
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"verbose\", \"nrounds\")"    
#> [3] "Modellist(verbose = 0, nrounds = 10)"
#> [4] "ModelRegression"                     
#> [5] "Model"