'resample' method for SplitCV objects

method that creates a ResamplePrediction. It contain the relevant result information of a resampling: the predictions made, the relevant ground truth values, and possibly the training tasks and trained model for each fold

resample(data, inducer, split_cv)

Arguments

data

A Dataset object

inducer

An Inducer Object

split_cv

The split method

Examples

cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
cars.split <- splt$cv(folds = 5)(cars.data)
resample(cars.data, xgb, cars.split)
#> $repeat_1_fold_1
#> $repeat_1_fold_1$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 38   68    19
#> 32   42    18
#> 7    18    10
#> 9    34    10
#> 43   64    20
#> 19   46    13
#> 50   85    25
#> 29   32    17
#> 6    10     9
#> 30   40    17
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 38   68    19
#> 32   42    18
#> 7    18    10
#> 9    34    10
#> 43   64    20
#> 19   46    13
#> 50   85    25
#> 29   32    17
#> 6    10     9
#> 30   40    17
#> 
#> $y
#>  [1] 19 18 10 10 20 13 25 17  9 17
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.3 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  37.437663
#>         2  29.078779
#> ---                 
#>         9   7.893999
#>        10   7.016990
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_1$predictions
#>    prediction truth
#> 1     9.38910     2
#> 2     9.38910    10
#> 3     9.38910     4
#> 4     9.38910    22
#> 5     9.38910    16
#> 8    24.95868    26
#> 10   24.95868    17
#> 11   24.95868    28
#> 12   38.61929    14
#> 13   38.61929    20
#> 14   38.61929    24
#> 15   38.61929    28
#> 16   38.61929    26
#> 17   38.61929    34
#> 18   38.61929    34
#> 20   38.61929    26
#> 21   38.61929    36
#> 22   38.61929    60
#> 23   38.61929    80
#> 24   35.86207    20
#> 25   35.86207    26
#> 26   35.86207    54
#> 27   35.86207    32
#> 28   35.86207    40
#> 31   35.86207    50
#> 33   40.33731    56
#> 34   40.33731    76
#> 35   40.33731    84
#> 36   62.77227    36
#> 37   62.77227    46
#> 39   62.77227    32
#> 40   62.77227    48
#> 41   62.77227    52
#> 42   62.77227    56
#> 44   62.77227    66
#> 45   69.38828    54
#> 46   69.38828    70
#> 47   69.38828    92
#> 48   69.38828    93
#> 49   69.38828   120
#> 
#> 
#> $repeat_1_fold_2
#> $repeat_1_fold_2$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 33   56    18
#> 5    16     8
#> 23   80    14
#> 25   26    15
#> 4    22     7
#> 12   14    12
#> 47   92    24
#> 1     2     4
#> 31   50    17
#> 15   28    12
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 33   56    18
#> 5    16     8
#> 23   80    14
#> 25   26    15
#> 4    22     7
#> 12   14    12
#> 47   92    24
#> 1     2     4
#> 31   50    17
#> 15   28    12
#> 
#> $y
#>  [1] 18  8 14 15  7 12 24  4 17 12
#> 
#> $model
#> ##### xgb.Booster
#> raw: 13.5 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  37.308791
#>         2  30.122782
#> ---                 
#>         9  10.027373
#>        10   8.694544
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_2$predictions
#>    prediction truth
#> 2    1.704688    10
#> 3   18.878887     4
#> 6   17.978888    10
#> 7   18.887407    18
#> 8   18.887407    26
#> 9   18.887407    34
#> 10  18.887407    17
#> 11  18.887407    28
#> 13  18.887407    20
#> 14  18.887407    24
#> 16  65.562981    26
#> 17  65.562981    34
#> 18  65.562981    34
#> 19  65.562981    46
#> 20  65.562981    26
#> 21  65.562981    36
#> 22  65.562981    60
#> 24  31.929056    20
#> 26  31.929056    54
#> 27  49.184036    32
#> 28  49.184036    40
#> 29  49.184036    32
#> 30  49.184036    40
#> 32  52.327648    42
#> 34  52.327648    76
#> 35  52.327648    84
#> 36  52.327648    36
#> 37  52.327648    46
#> 38  52.327648    68
#> 39  52.327648    32
#> 40  52.327648    48
#> 41  52.327648    52
#> 42  52.327648    56
#> 43  52.327648    64
#> 44  72.569427    66
#> 45  72.569427    54
#> 46  72.569427    70
#> 48  72.569427    93
#> 49  72.569427   120
#> 50  72.569427    85
#> 
#> 
#> $repeat_1_fold_3
#> $repeat_1_fold_3$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 49  120    24
#> 11   28    11
#> 24   20    15
#> 46   70    24
#> 2    10     4
#> 37   46    19
#> 16   26    13
#> 3     4     7
#> 45   54    23
#> 41   52    20
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 49  120    24
#> 11   28    11
#> 24   20    15
#> 46   70    24
#> 2    10     4
#> 37   46    19
#> 16   26    13
#> 3     4     7
#> 45   54    23
#> 41   52    20
#> 
#> $y
#>  [1] 24 11 15 24  4 19 13  7 23 20
#> 
#> $model
#> ##### xgb.Booster
#> raw: 13.1 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1   42.15338
#>         2   34.14140
#> ---                 
#>         9   13.07155
#>        10   12.43242
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_3$predictions
#>    prediction truth
#> 1    8.223087     2
#> 4    4.146549    22
#> 5    4.146549    16
#> 6   24.755972    10
#> 7   24.755972    18
#> 8   24.755972    26
#> 9   24.755972    34
#> 10  24.755972    17
#> 12  24.755972    14
#> 13  24.755972    20
#> 14  24.755972    24
#> 15  24.755972    28
#> 17  24.755972    34
#> 18  24.755972    34
#> 19  24.755972    46
#> 20  21.607246    26
#> 21  21.607246    36
#> 22  21.607246    60
#> 23  21.607246    80
#> 25  21.607246    26
#> 26  21.607246    54
#> 27  21.607246    32
#> 28  21.607246    40
#> 29  44.534454    32
#> 30  44.534454    40
#> 31  44.534454    50
#> 32  44.534454    42
#> 33  44.534454    56
#> 34  44.534454    76
#> 35  44.534454    84
#> 36  44.534454    36
#> 38  44.534454    68
#> 39  47.525826    32
#> 40  47.525826    48
#> 42  47.525826    56
#> 43  47.525826    64
#> 44  51.176704    66
#> 47  83.871628    92
#> 48  83.871628    93
#> 50  83.871628    85
#> 
#> 
#> $repeat_1_fold_4
#> $repeat_1_fold_4$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 13   20    12
#> 48   93    24
#> 34   76    18
#> 39   32    20
#> 35   84    18
#> 17   34    13
#> 40   48    20
#> 22   60    14
#> 28   40    16
#> 27   32    16
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 13   20    12
#> 48   93    24
#> 34   76    18
#> 39   32    20
#> 35   84    18
#> 17   34    13
#> 40   48    20
#> 22   60    14
#> 28   40    16
#> 27   32    16
#> 
#> $y
#>  [1] 12 24 18 20 18 13 20 14 16 16
#> 
#> $model
#> ##### xgb.Booster
#> raw: 13 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1   43.90428
#>         2   35.04533
#> ---                 
#>         9   10.88510
#>        10    9.39124
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_4$predictions
#>    prediction truth
#> 1    19.10866     2
#> 2    19.10866    10
#> 3    19.10866     4
#> 4    19.10866    22
#> 5    19.10866    16
#> 6    19.10866    10
#> 7    19.10866    18
#> 8    19.10866    26
#> 9    19.10866    34
#> 10   19.10866    17
#> 11   19.10866    28
#> 12   19.10866    14
#> 14   19.10866    24
#> 15   19.10866    28
#> 16   32.30879    26
#> 18   32.30879    34
#> 19   32.30879    46
#> 20   50.11182    26
#> 21   50.11182    36
#> 23   50.11182    80
#> 24   34.89761    20
#> 25   34.89761    26
#> 26   34.89761    54
#> 29   70.14961    32
#> 30   70.14961    40
#> 31   70.14961    50
#> 32   70.14961    42
#> 33   70.14961    56
#> 36   43.25682    36
#> 37   43.25682    46
#> 38   43.25682    68
#> 41   43.25682    52
#> 42   43.25682    56
#> 43   43.25682    64
#> 44   73.74335    66
#> 45   73.74335    54
#> 46   73.74335    70
#> 47   73.74335    92
#> 49   73.74335   120
#> 50   73.74335    85
#> 
#> 
#> $repeat_1_fold_5
#> $repeat_1_fold_5$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 18   34    13
#> 36   36    19
#> 8    26    10
#> 20   26    14
#> 42   56    20
#> 10   17    11
#> 44   66    22
#> 14   24    12
#> 21   36    14
#> 26   54    15
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 18   34    13
#> 36   36    19
#> 8    26    10
#> 20   26    14
#> 42   56    20
#> 10   17    11
#> 44   66    22
#> 14   24    12
#> 21   36    14
#> 26   54    15
#> 
#> $y
#>  [1] 13 19 10 14 20 11 22 12 14 15
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.6 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  30.639935
#>         2  23.698446
#> ---                 
#>         9   6.106534
#>        10   5.321504
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_5$predictions
#>    prediction truth
#> 1    22.49945     2
#> 2    22.49945    10
#> 3    22.49945     4
#> 4    22.49945    22
#> 5    22.49945    16
#> 6    22.49945    10
#> 7    22.49945    18
#> 9    22.49945    34
#> 11   19.02658    28
#> 12   22.85428    14
#> 13   22.85428    20
#> 15   22.85428    28
#> 16   30.75776    26
#> 17   30.75776    34
#> 19   30.75776    46
#> 22   30.34848    60
#> 23   30.34848    80
#> 24   46.62649    20
#> 25   46.62649    26
#> 27   46.62649    32
#> 28   46.62649    40
#> 29   37.32992    32
#> 30   37.32992    40
#> 31   37.32992    50
#> 32   37.32992    42
#> 33   37.32992    56
#> 34   37.32992    76
#> 35   37.32992    84
#> 37   37.32992    46
#> 38   37.32992    68
#> 39   52.77045    32
#> 40   52.77045    48
#> 41   52.77045    52
#> 43   52.77045    64
#> 45   54.27045    54
#> 46   54.27045    70
#> 47   54.27045    92
#> 48   54.27045    93
#> 49   54.27045   120
#> 50   54.27045    85
#> 
#>