'resample' method for SplitCV objects

method that creates a ResamplePrediction. It contain the relevant result information of a resampling: the predictions made, the relevant ground truth values, and possibly the training tasks and trained model for each fold

resample(data, inducer, split_cv)

Arguments

data

A Dataset object

inducer

An Inducer Object

split_cv

The split method

Examples

cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
cars.split <- splt$cv(folds = 5)(cars.data)
resample(cars.data, xgb, cars.split)
#> $repeat_1_fold_1
#> $repeat_1_fold_1$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 5    16     8
#> 23   80    14
#> 38   68    19
#> 4    22     7
#> 12   14    12
#> 13   20    12
#> 1     2     4
#> 49  120    24
#> 42   56    20
#> 2    10     4
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 5    16     8
#> 23   80    14
#> 38   68    19
#> 4    22     7
#> 12   14    12
#> 13   20    12
#> 1     2     4
#> 49  120    24
#> 42   56    20
#> 2    10     4
#> 
#> $y
#>  [1]  8 14 19  7 12 12  4 24 20  4
#> 
#> $model
#> ##### xgb.Booster
#> raw: 13.4 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  42.551439
#>         2  33.866319
#> ---                 
#>         9  10.368368
#>        10   8.889586
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_1$predictions
#>    prediction truth
#> 3    18.15164     4
#> 6    15.74826    10
#> 7    15.97616    18
#> 8    15.97616    26
#> 9    15.97616    34
#> 10   15.97616    17
#> 11   15.97616    28
#> 14   15.97616    24
#> 15   15.97616    28
#> 16   70.07172    26
#> 17   70.07172    34
#> 18   70.07172    34
#> 19   70.07172    46
#> 20   70.07172    26
#> 21   70.07172    36
#> 22   70.07172    60
#> 24   70.07172    20
#> 25   70.07172    26
#> 26   70.07172    54
#> 27   70.07172    32
#> 28   70.07172    40
#> 29   65.37308    32
#> 30   65.37308    40
#> 31   65.37308    50
#> 32   65.37308    42
#> 33   65.37308    56
#> 34   65.37308    76
#> 35   65.37308    84
#> 36   65.37308    36
#> 37   65.37308    46
#> 39   55.55965    32
#> 40   55.55965    48
#> 41   55.55965    52
#> 43   55.55965    64
#> 44   95.16223    66
#> 45   95.16223    54
#> 46   95.16223    70
#> 47   95.16223    92
#> 48   95.16223    93
#> 50   95.16223    85
#> 
#> 
#> $repeat_1_fold_2
#> $repeat_1_fold_2$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 11   28    11
#> 24   20    15
#> 25   26    15
#> 43   64    20
#> 31   50    17
#> 16   26    13
#> 3     4     7
#> 26   54    15
#> 44   66    22
#> 41   52    20
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 11   28    11
#> 24   20    15
#> 25   26    15
#> 43   64    20
#> 31   50    17
#> 16   26    13
#> 3     4     7
#> 26   54    15
#> 44   66    22
#> 41   52    20
#> 
#> $y
#>  [1] 11 15 15 20 17 13  7 15 22 20
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.1 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  33.435697
#>         2  26.106653
#> ---                 
#>         9   9.681809
#>        10   9.339508
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_2$predictions
#>    prediction truth
#> 1    3.617297     2
#> 2    3.617297    10
#> 4    3.617297    22
#> 5    3.617297    16
#> 6   26.555292    10
#> 7   26.555292    18
#> 8   26.555292    26
#> 9   26.555292    34
#> 10  26.555292    17
#> 12  26.290144    14
#> 13  26.290144    20
#> 14  26.290144    24
#> 15  26.290144    28
#> 17  26.290144    34
#> 18  26.290144    34
#> 19  26.290144    46
#> 20  31.797541    26
#> 21  31.797541    36
#> 22  31.797541    60
#> 23  31.797541    80
#> 27  48.042599    32
#> 28  48.042599    40
#> 29  48.042599    32
#> 30  48.042599    40
#> 32  48.042599    42
#> 33  48.042599    56
#> 34  48.042599    76
#> 35  48.042599    84
#> 36  55.321716    36
#> 37  55.321716    46
#> 38  55.321716    68
#> 39  55.321716    32
#> 40  55.321716    48
#> 42  55.321716    56
#> 45  55.321716    54
#> 46  55.321716    70
#> 47  55.321716    92
#> 48  55.321716    93
#> 49  55.321716   120
#> 50  55.321716    85
#> 
#> 
#> $repeat_1_fold_3
#> $repeat_1_fold_3$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 33   56    18
#> 34   76    18
#> 35   84    18
#> 37   46    19
#> 17   34    13
#> 46   70    24
#> 47   92    24
#> 48   93    24
#> 20   26    14
#> 50   85    25
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 33   56    18
#> 34   76    18
#> 35   84    18
#> 37   46    19
#> 17   34    13
#> 46   70    24
#> 47   92    24
#> 48   93    24
#> 20   26    14
#> 50   85    25
#> 
#> $y
#>  [1] 18 18 18 19 13 24 24 24 14 25
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.3 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  52.159711
#>         2  39.615670
#> ---                 
#>         9  10.634539
#>        10   9.889692
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_3$predictions
#>    prediction truth
#> 1    30.21178     2
#> 2    30.21178    10
#> 3    30.21178     4
#> 4    30.21178    22
#> 5    30.21178    16
#> 6    30.21178    10
#> 7    30.21178    18
#> 8    30.21178    26
#> 9    30.21178    34
#> 10   30.21178    17
#> 11   30.21178    28
#> 12   30.21178    14
#> 13   30.21178    20
#> 14   30.21178    24
#> 15   30.21178    28
#> 16   30.21178    26
#> 18   30.21178    34
#> 19   30.21178    46
#> 21   26.38783    36
#> 22   26.38783    60
#> 23   26.38783    80
#> 24   26.38783    20
#> 25   26.38783    26
#> 26   26.38783    54
#> 27   67.51583    32
#> 28   67.51583    40
#> 29   67.51583    32
#> 30   67.51583    40
#> 31   67.51583    50
#> 32   67.51583    42
#> 36   48.36531    36
#> 38   48.36531    68
#> 39   48.36531    32
#> 40   48.36531    48
#> 41   48.36531    52
#> 42   48.36531    56
#> 43   48.36531    64
#> 44   79.00748    66
#> 45   79.00748    54
#> 49   79.00748   120
#> 
#> 
#> $repeat_1_fold_4
#> $repeat_1_fold_4$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 36   36    19
#> 8    26    10
#> 18   34    13
#> 30   40    17
#> 10   17    11
#> 21   36    14
#> 28   40    16
#> 15   28    12
#> 22   60    14
#> 14   24    12
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 36   36    19
#> 8    26    10
#> 18   34    13
#> 30   40    17
#> 10   17    11
#> 21   36    14
#> 28   40    16
#> 15   28    12
#> 22   60    14
#> 14   24    12
#> 
#> $y
#>  [1] 19 10 13 17 11 14 16 12 14 12
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.1 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  26.863096
#>         2  20.592494
#> ---                 
#>         9   6.652381
#>        10   6.252455
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_4$predictions
#>    prediction truth
#> 1    22.44849     2
#> 2    22.44849    10
#> 3    22.44849     4
#> 4    22.44849    22
#> 5    22.44849    16
#> 6    22.44849    10
#> 7    22.44849    18
#> 9    22.44849    34
#> 11   18.97561    28
#> 12   24.59251    14
#> 13   24.59251    20
#> 16   32.68624    26
#> 17   32.68624    34
#> 19   32.68624    46
#> 20   42.25411    26
#> 23   42.25411    80
#> 24   38.36000    20
#> 25   38.36000    26
#> 26   38.36000    54
#> 27   38.36000    32
#> 29   38.36000    32
#> 31   38.36000    50
#> 32   34.85198    42
#> 33   34.85198    56
#> 34   34.85198    76
#> 35   34.85198    84
#> 37   34.85198    46
#> 38   34.85198    68
#> 39   34.85198    32
#> 40   34.85198    48
#> 41   34.85198    52
#> 42   34.85198    56
#> 43   34.85198    64
#> 44   34.85198    66
#> 45   34.85198    54
#> 46   34.85198    70
#> 47   34.85198    92
#> 48   34.85198    93
#> 49   34.85198   120
#> 50   34.85198    85
#> 
#> 
#> $repeat_1_fold_5
#> $repeat_1_fold_5$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"       
#> 
#> $data
#> $data
#>    dist speed
#> 7    18    10
#> 9    34    10
#> 45   54    23
#> 19   46    13
#> 32   42    18
#> 40   48    20
#> 6    10     9
#> 27   32    16
#> 39   32    20
#> 29   32    17
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 7    18    10
#> 9    34    10
#> 45   54    23
#> 19   46    13
#> 32   42    18
#> 40   48    20
#> 6    10     9
#> 27   32    16
#> 39   32    20
#> 29   32    17
#> 
#> $y
#>  [1] 10 10 23 13 18 20  9 16 20 17
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.5 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  27.584453
#>         2  21.129645
#> ---                 
#>         9   7.016015
#>        10   6.520402
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_5$predictions
#>    prediction truth
#> 1    9.347319     2
#> 2    9.347319    10
#> 3    9.347319     4
#> 4    9.347319    22
#> 5    9.347319    16
#> 8   24.815767    26
#> 10  24.815767    17
#> 11  24.815767    28
#> 12  39.433372    14
#> 13  39.433372    20
#> 14  39.433372    24
#> 15  39.433372    28
#> 16  39.433372    26
#> 17  39.433372    34
#> 18  39.433372    34
#> 20  39.433372    26
#> 21  39.433372    36
#> 22  39.433372    60
#> 23  39.433372    80
#> 24  31.884920    20
#> 25  31.884920    26
#> 26  31.884920    54
#> 28  31.884920    40
#> 30  31.884920    40
#> 31  31.884920    50
#> 33  38.897491    56
#> 34  38.897491    76
#> 35  38.897491    84
#> 36  38.897491    36
#> 37  38.897491    46
#> 38  38.897491    68
#> 41  38.897491    52
#> 42  38.897491    56
#> 43  38.897491    64
#> 44  43.477551    66
#> 46  43.477551    70
#> 47  43.477551    92
#> 48  43.477551    93
#> 49  43.477551   120
#> 50  43.477551    85
#> 
#>