'resample' method for SplitCV objects

method that creates a ResamplePrediction. It contain the relevant result information of a resampling: the predictions made, the relevant ground truth values, and possibly the training tasks and trained model for each fold

resample(data, inducer, split_cv)

Arguments

data

A Dataset object

inducer

An Inducer Object

split_cv

The split method

Examples

cars.data <- Dataset(data = cars, target = "dist")
xgb <- InducerConstructer(configuration = list(nrounds = 10, verbose = 0), method = "XGBoost")
cars.split <- splt$cv(folds = 5)(cars.data)
resample(cars.data, xgb, cars.split)
#> $repeat_1_fold_1
#> $repeat_1_fold_1$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
#> 
#> $data
#> $data
#>    dist speed
#> 8    26    10
#> 46   70    24
#> 4    22     7
#> 12   14    12
#> 16   26    13
#> 3     4     7
#> 24   20    15
#> 26   54    15
#> 5    16     8
#> 14   24    12
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 8    26    10
#> 46   70    24
#> 4    22     7
#> 12   14    12
#> 16   26    13
#> 3     4     7
#> 24   20    15
#> 26   54    15
#> 5    16     8
#> 14   24    12
#> 
#> $y
#>  [1] 10 24  7 12 13  7 15 15  8 12
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.4 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1   26.55112
#>         2   21.95732
#> ---                 
#>         9   10.59507
#>        10   10.14276
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_1$predictions
#>    prediction truth
#> 1    12.46964     2
#> 2    12.46964    10
#> 6    21.13803    10
#> 7    21.13803    18
#> 9    21.13803    34
#> 10   19.17320    17
#> 11   19.17320    28
#> 13   19.17320    20
#> 15   19.17320    28
#> 17   25.23151    34
#> 18   25.23151    34
#> 19   25.23151    46
#> 20   34.76458    26
#> 21   34.76458    36
#> 22   34.76458    60
#> 23   34.76458    80
#> 25   34.76458    26
#> 27   34.76458    32
#> 28   34.76458    40
#> 29   34.76458    32
#> 30   34.76458    40
#> 31   34.76458    50
#> 32   34.76458    42
#> 33   34.76458    56
#> 34   34.76458    76
#> 35   34.76458    84
#> 36   34.76458    36
#> 37   34.76458    46
#> 38   34.76458    68
#> 39   55.74221    32
#> 40   55.74221    48
#> 41   55.74221    52
#> 42   55.74221    56
#> 43   55.74221    64
#> 44   55.74221    66
#> 45   55.74221    54
#> 47   55.74221    92
#> 48   55.74221    93
#> 49   55.74221   120
#> 50   55.74221    85
#> 
#> 
#> $repeat_1_fold_2
#> $repeat_1_fold_2$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
#> 
#> $data
#> $data
#>    dist speed
#> 9    34    10
#> 25   26    15
#> 34   76    18
#> 17   34    13
#> 7    18    10
#> 29   32    17
#> 11   28    11
#> 28   40    16
#> 44   66    22
#> 15   28    12
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 9    34    10
#> 25   26    15
#> 34   76    18
#> 17   34    13
#> 7    18    10
#> 29   32    17
#> 11   28    11
#> 28   40    16
#> 44   66    22
#> 15   28    12
#> 
#> $y
#>  [1] 10 15 18 13 10 17 11 16 22 12
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.1 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  32.497950
#>         2  25.660377
#> ---                 
#>         9   6.963565
#>        10   6.174533
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_2$predictions
#>    prediction truth
#> 1    25.50453     2
#> 2    25.50453    10
#> 3    25.50453     4
#> 4    25.50453    22
#> 5    25.50453    16
#> 6    25.50453    10
#> 8    25.50453    26
#> 10   27.17980    17
#> 12   27.17980    14
#> 13   27.17980    20
#> 14   27.17980    24
#> 16   29.43847    26
#> 18   29.43847    34
#> 19   29.43847    46
#> 20   28.23847    26
#> 21   28.23847    36
#> 22   28.23847    60
#> 23   28.23847    80
#> 24   28.23847    20
#> 26   28.23847    54
#> 27   34.11388    32
#> 30   31.21999    40
#> 31   31.21999    50
#> 32   62.62328    42
#> 33   62.62328    56
#> 35   62.62328    84
#> 36   62.62328    36
#> 37   62.62328    46
#> 38   62.62328    68
#> 39   62.62328    32
#> 40   62.62328    48
#> 41   62.62328    52
#> 42   62.62328    56
#> 43   62.62328    64
#> 45   62.62328    54
#> 46   62.62328    70
#> 47   62.62328    92
#> 48   62.62328    93
#> 49   62.62328   120
#> 50   62.62328    85
#> 
#> 
#> $repeat_1_fold_3
#> $repeat_1_fold_3$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
#> 
#> $data
#> $data
#>    dist speed
#> 41   52    20
#> 35   84    18
#> 42   56    20
#> 10   17    11
#> 21   36    14
#> 39   32    20
#> 33   56    18
#> 27   32    16
#> 31   50    17
#> 48   93    24
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 41   52    20
#> 35   84    18
#> 42   56    20
#> 10   17    11
#> 21   36    14
#> 39   32    20
#> 33   56    18
#> 27   32    16
#> 31   50    17
#> 48   93    24
#> 
#> $y
#>  [1] 20 18 20 11 14 20 18 16 17 24
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.5 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1   42.36643
#>         2   33.21162
#> ---                 
#>         9   11.94578
#>        10   11.03045
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_3$predictions
#>    prediction truth
#> 1    16.37477     2
#> 2    16.37477    10
#> 3    16.37477     4
#> 4    16.37477    22
#> 5    16.37477    16
#> 6    16.37477    10
#> 7    16.37477    18
#> 8    16.37477    26
#> 9    16.37477    34
#> 11   16.37477    28
#> 12   16.37477    14
#> 13   16.37477    20
#> 14   16.37477    24
#> 15   16.37477    28
#> 16   31.79494    26
#> 17   31.79494    34
#> 18   31.79494    34
#> 19   31.79494    46
#> 20   31.79494    26
#> 22   31.79494    60
#> 23   31.79494    80
#> 24   31.79494    20
#> 25   31.79494    26
#> 26   31.79494    54
#> 28   31.79494    40
#> 29   48.12875    32
#> 30   48.12875    40
#> 32   62.64678    42
#> 34   62.64678    76
#> 36   45.77658    36
#> 37   45.77658    46
#> 38   45.77658    68
#> 40   45.77658    48
#> 43   45.77658    64
#> 44   73.97610    66
#> 45   73.97610    54
#> 46   73.97610    70
#> 47   73.97610    92
#> 49   73.97610   120
#> 50   73.97610    85
#> 
#> 
#> $repeat_1_fold_4
#> $repeat_1_fold_4$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
#> 
#> $data
#> $data
#>    dist speed
#> 38   68    19
#> 18   34    13
#> 30   40    17
#> 43   64    20
#> 40   48    20
#> 6    10     9
#> 50   85    25
#> 2    10     4
#> 47   92    24
#> 49  120    24
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 38   68    19
#> 18   34    13
#> 30   40    17
#> 43   64    20
#> 40   48    20
#> 6    10     9
#> 50   85    25
#> 2    10     4
#> 47   92    24
#> 49  120    24
#> 
#> $y
#>  [1] 19 13 17 20 20  9 25  4 24 24
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.9 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1  50.689622
#>         2  39.803006
#> ---                 
#>         9  10.813787
#>        10   9.743679
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_4$predictions
#>    prediction truth
#> 1    9.245864     2
#> 3    9.245864     4
#> 4    9.245864    22
#> 5    9.245864    16
#> 7    9.245864    18
#> 8    9.245864    26
#> 9    9.245864    34
#> 10  30.608963    17
#> 11  30.608963    28
#> 12  30.608963    14
#> 13  30.608963    20
#> 14  30.608963    24
#> 15  30.608963    28
#> 16  30.608963    26
#> 17  30.608963    34
#> 19  30.608963    46
#> 20  30.608963    26
#> 21  30.608963    36
#> 22  30.608963    60
#> 23  30.608963    80
#> 24  39.867992    20
#> 25  39.867992    26
#> 26  39.867992    54
#> 27  39.867992    32
#> 28  39.867992    40
#> 29  39.867992    32
#> 31  39.867992    50
#> 32  58.664894    42
#> 33  58.664894    56
#> 34  58.664894    76
#> 35  58.664894    84
#> 36  58.664894    36
#> 37  58.664894    46
#> 39  54.361259    32
#> 41  54.361259    52
#> 42  54.361259    56
#> 44  93.709351    66
#> 45  93.709351    54
#> 46  93.709351    70
#> 48  93.709351    93
#> 
#> 
#> $repeat_1_fold_5
#> $repeat_1_fold_5$model
#> $inducer
#> $method
#> [1] "XGBoost"
#> 
#> $hyperparameters
#> [1] "nrounds" "verbose"
#> 
#> $configuration
#> $configuration$nrounds
#> [1] 10
#> 
#> $configuration$verbose
#> [1] 0
#> 
#> 
#> attr(,"class")
#> [1] "InducerXGBoost" "Inducer"        "function"      
#> 
#> $data
#> $data
#>    dist speed
#> 45   54    23
#> 23   80    14
#> 19   46    13
#> 13   20    12
#> 1     2     4
#> 32   42    18
#> 22   60    14
#> 37   46    19
#> 36   36    19
#> 20   26    14
#> 
#> $target
#> [1] "dist"
#> 
#> $type
#> [1] "regression"
#> 
#> $name
#> [1] "cars"
#> 
#> attr(,"class")
#> [1] "DatasetRegression" "Dataset"          
#> 
#> $X
#>    dist speed
#> 45   54    23
#> 23   80    14
#> 19   46    13
#> 13   20    12
#> 1     2     4
#> 32   42    18
#> 22   60    14
#> 37   46    19
#> 36   36    19
#> 20   26    14
#> 
#> $y
#>  [1] 23 14 13 12  4 18 14 19 19 14
#> 
#> $model
#> ##### xgb.Booster
#> raw: 12.7 Kb 
#> call:
#>   xgb.train(params = params, data = dtrain, nrounds = nrounds, 
#>     watchlist = watchlist, verbose = verbose, print_every_n = print_every_n, 
#>     early_stopping_rounds = early_stopping_rounds, maximize = maximize, 
#>     save_period = save_period, save_name = save_name, xgb_model = xgb_model, 
#>     callbacks = callbacks, objective = "reg:squarederror")
#> params (as set within xgb.train):
#>   objective = "reg:squarederror", validate_parameters = "TRUE"
#> xgb.attributes:
#>   niter
#> callbacks:
#>   cb.evaluation.log()
#> # of features: 1 
#> niter: 10
#> nfeatures : 1 
#> evaluation_log:
#>      iter train_rmse
#>     <num>      <num>
#>         1   34.87649
#>         2   27.30836
#> ---                 
#>         9   13.18574
#>        10   12.91653
#> 
#> $task
#> [1] "Regression"
#> 
#> attr(,"class")
#> [1] "ModelXGBoost"                        
#> [2] "Modelc(\"nrounds\", \"verbose\")"    
#> [3] "Modellist(nrounds = 10, verbose = 0)"
#> [4] "ModelRegression"                     
#> [5] "Model"                               
#> 
#> $repeat_1_fold_5$predictions
#>    prediction truth
#> 2    1.704688    10
#> 3    1.704688     4
#> 4    1.704688    22
#> 5   18.255913    16
#> 6   18.255913    10
#> 7   18.255913    18
#> 8   18.255913    26
#> 9   18.255913    34
#> 10  18.255913    17
#> 11  18.255913    28
#> 12  18.255913    14
#> 14  18.255913    24
#> 15  18.255913    28
#> 16  44.137085    26
#> 17  44.137085    34
#> 18  44.137085    34
#> 21  50.492680    36
#> 24  50.492680    20
#> 25  50.492680    26
#> 26  50.492680    54
#> 27  40.505444    32
#> 28  40.505444    40
#> 29  40.505444    32
#> 30  40.505444    40
#> 31  40.505444    50
#> 33  40.505444    56
#> 34  40.505444    76
#> 35  40.505444    84
#> 38  40.505444    68
#> 39  40.505444    32
#> 40  40.505444    48
#> 41  40.505444    52
#> 42  40.505444    56
#> 43  40.505444    64
#> 44  47.065842    66
#> 46  47.065842    70
#> 47  47.065842    92
#> 48  47.065842    93
#> 49  47.065842   120
#> 50  47.065842    85
#> 
#>